+++ ecomm/views/__init__.py
@@ -1,6 +1,7 @@
-from .auth import index, register, login_user, user_logout, sell_product, list_products
+from .auth import index, register, login_user, user_logout, list_products
 from .userSettings import userSettings
 from .productDetail import productDetail
 from .searchProduct import search, choose
 from .userSettings import userSettings
 from .shoppingCart import cart_items_list, deleteOrderItem
+from .sellProduct import sell_product
